+++ go.mod
@@ -3,24 +3,6 @@ module wombat
 go 1.25.1
 
 require (
-	github.com/andygrunwald/go-jira v1.16.0
-	github.com/google/cel-go v0.26.1
 	gopkg.in/yaml.v2 v2.4.0
 	mvdan.cc/sh/v3 v3.12.0
-)
-
-require (
-	cel.dev/expr v0.24.0 // indirect
-	github.com/antlr4-go/antlr/v4 v4.13.1 // indirect
-	github.com/fatih/structs v1.1.0 // indirect
-	github.com/golang-jwt/jwt/v4 v4.5.2 // indirect
-	github.com/google/go-querystring v1.1.0 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	github.com/stoewer/go-strcase v1.3.1 // indirect
-	github.com/trivago/tgo v1.0.7 // indirect
-	golang.org/x/exp v0.0.0-20250911091902-df9299821621 // indirect
-	golang.org/x/text v0.29.0 // indirect
-	google.golang.org/genproto/googleapis/api v0.0.0-20250908214217-97024824d090 // indirect
-	google.golang.org/genproto/googleapis/rpc v0.0.0-20250908214217-97024824d090 // indirect
-	google.golang.org/protobuf v1.36.9 // indirect
 )
