+++ plugins/jira/go.mod
@@ -1,0 +1,13 @@
+module jira
+
+go 1.25.1
+
+require github.com/andygrunwald/go-jira v1.16.0
+
+require (
+	github.com/fatih/structs v1.1.0 // indirect
+	github.com/golang-jwt/jwt/v4 v4.5.2 // indirect
+	github.com/google/go-querystring v1.1.0 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
+	github.com/trivago/tgo v1.0.7 // indirect
+)
